IoT/아두이노-라즈베리파이/django-camera/requirements.txt
absl-py==0.9.0
asgiref==3.2.10
astor==0.8.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-cors-headers==3.4.0
django-redis==4.12.1
djangorestframework==3.11.0
gast==0.2.2
google-auth==1.20.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.30.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
Markdown==3.2.2
numpy==1.19.1
oauthlib==3.1.0
opt-einsum==3.3.0
picamera==1.13
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.12.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyserial==3.4
pytz==2020.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scipy==1.4.1
six==1.15.0
sqlparse==0.3.1
tensorboard==2.1.1
tensorflow @ file:///home/pi/s03p13b109/IoT/%EC%95%84%EB%91%90%EC%9D%B4%EB%85%B8-%EB%9D%BC%EC%A6%88%EB%B2%A0%EB%A6%AC%ED%8C%8C%EC%9D%B4/django-camera/tensorflow-2.1.0-cp37-cp37m-linux_armv7l.whl
tensorflow-estimator==2.1.0
termcolor==1.1.0
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
